requirements.txt
paho-mqtt
psycopg2-binary
python-dotenv
pymysql
mysql-connector-python
flask
flask-socketio
python-socketio
requests
supabase